+++ Sources/EV3/EV3DrawLineCommand.swift
@@ -26,7 +26,7 @@ public struct EV3DrawLineCommand: EV3DirectCommand {
 		return EV3GenericResponse.self
 	}
 
-	public func payloadDataWithGlobalOffset(offset: UInt8) -> NSData {
+	public func payloadDataWithGlobalOffset(offset: UInt16) -> NSData {
 		let mutableData = NSMutableData()
 		mutableData.appendUInt8(EV3OpCode.UIDraw.rawValue)
 		mutableData.appendUInt8(EV3UIDrawOpSubcode.Line.rawValue)
